T6895: Merge the hsflowd-based sFlow and uacctd-based sFlow
