+++ interface-definitions/include/version/flow-accounting-version.xml.i
@@ -1,3 +1,3 @@
 <!-- include start from include/version/flow-accounting-version.xml.i -->
-<syntaxVersion component='flow-accounting' version='1'></syntaxVersion>
+<syntaxVersion component='flow-accounting' version='2'></syntaxVersion>
 <!-- include end -->
