+++ interface-definitions/system_flow-accounting.xml.in
@@ -362,73 +362,6 @@
               </node>
             </children>
           </node>
-          <node name="sflow">
-            <properties>
-              <help>sFlow settings</help>
-            </properties>
-            <children>
-              <leafNode name="agent-address">
-                <properties>
-                  <help>sFlow agent IPv4 address</help>
-                  <completionHelp>
-                    <list>auto</list>
-                    <script>${vyos_completion_dir}/list_local_ips.sh --ipv4</script>
-                  </completionHelp>
-                  <valueHelp>
-                    <format>ipv4</format>
-                    <description>sFlow IPv4 agent address</description>
-                  </valueHelp>
-                  <constraint>
-                    <validator name="ipv4-address"/>
-                  </constraint>
-                </properties>
-              </leafNode>
-              <leafNode name="sampling-rate">
-                <properties>
-                  <help>sFlow sampling-rate</help>
-                  <valueHelp>
-                    <format>u32</format>
-                    <description>Sampling rate (1 in N packets)</description>
-                  </valueHelp>
-                  <constraint>
-                   <validator name="numeric" argument="--range 0-4294967295"/>
-                  </constraint>
-                </properties>
-              </leafNode>
-              <tagNode name="server">
-                <properties>
-                  <help>sFlow destination server</help>
-                  <valueHelp>
-                    <format>ipv4</format>
-                    <description>IPv4 server to export sFlow</description>
-                  </valueHelp>
-                  <valueHelp>
-                    <format>ipv6</format>
-                    <description>IPv6 server to export sFlow</description>
-                  </valueHelp>
-                  <constraint>
-                    <validator name="ip-address"/>
-                  </constraint>
-                </properties>
-                <children>
-                  <leafNode name="port">
-                    <properties>
-                      <help>sFlow port number</help>
-                      <valueHelp>
-                        <format>u32:1025-65535</format>
-                        <description>sFlow port number</description>
-                      </valueHelp>
-                      <constraint>
-                       <validator name="numeric" argument="--range 1025-65535"/>
-                      </constraint>
-                    </properties>
-                    <defaultValue>6343</defaultValue>
-                  </leafNode>
-                </children>
-              </tagNode>
-              #include <include/source-address-ipv4-ipv6.xml.i>
-            </children>
-          </node>
           #include <include/interface/vrf.xml.i>
         </children>
       </node>
