+++ smoketest/config-tests/bgp-big-as-cloud
@@ -836,7 +836,6 @@ set system flow-accounting interface 'eth0.4089'
 set system flow-accounting netflow engine-id '1'
 set system flow-accounting netflow server 192.0.2.55 port '2055'
 set system flow-accounting netflow version '9'
-set system flow-accounting sflow server 1.2.3.4 port '1234'
 set system flow-accounting syslog-facility 'daemon'
 set system host-name 'vyos'
 set system login user vyos authentication encrypted-password '$6$2Ta6TWHd/U$NmrX0x9kexCimeOcYK1MfhMpITF9ELxHcaBU/znBq.X2ukQOj61fVI2UYP/xBzP4QtiTcdkgs7WOQMHWsRymO/'
@@ -845,6 +844,9 @@ set system name-server '2001:db8::1'
 set system name-server '2001:db8::2'
 set system name-server '192.0.2.1'
 set system name-server '192.0.2.2'
+set system sflow interface 'eth0.4088'
+set system sflow interface 'eth0.4089'
+set system sflow server 1.2.3.4 port '1234'
 set system syslog global facility all level 'all'
 set system syslog global preserve-fqdn
 set system time-zone 'Europe/Zurich'
